hardware: Add very basic programmable IO
This peripheral allows to send commands to a state machine to controll
IOs. Currently: HIGH, LOW, WAIT and READ are supported.

HIGH: Sets direction to out and level to high.
LOW: Sets direction to out and level to low.
WAIT: Waits x counter cycles.
READ: Sets direction to in, waits some cycles and reads the level.

Wait counter is configurable and number of counts is given.
Delay cycles before READ is configurable.

Signed-off-by: Daniel Schultz <daniel.schultz@aesc-silicon.de>